librtas: PCI hotplug support vs little endian
The goal of this patch is to enable a little endian process to interact
with RTAS which is big endian. The rationale is that librtas assumes
to be passed cpu endian arguments and do the necessary conversions. The
reasons for this choice are:
- many arguments are owned by librtas (status, RMO buffer addresses...)
- nearly all arguments are being used in cpu endian order by librtas:
  if they were assumed to be big endian, we would have to byte swap
  anyway, resulting in an even bigger patch...

Of course, all rules have an exception: the RMO buffer content is not
seen by librtas, it is up to the caller (drmgr) to do the conversions.

It seems that recent kernels don't have the needed procfs files anymore,
so this patch only cares about the rtas() syscall based implementation.

As requested by Benjamin Herrenschmidt, the byte swapping should be
performed in userland AND be careful about corner cases where non
32-bit values are being passed:

http://patchwork.ozlabs.org/patch/327089/

We cannot boldly convert all arguments in the sc_rtas_call() function.

For an example, let's look at sc_suspend_me() where we pass a 64 bit value
split across two arguments. If streamid is ABCDEFGHIJKLMNOP, then it will
be reassembled by the kernel as IJKLMNOPABCDEFGH before being passed to the
host...

As a consequence, we can only convert well known parameters in the
sc_rtas_call() function:
- token
- number of input arguments
- number of return arguments
- return status (always the first return argument)

All other values must be converted in the caller. Please note that we
introduce intermediary variables for return values as we do not want
big endian values to be written in memory where the code assumes they
are cpu endian.

As a starter, we focus on the RTAS calls involved in PCI hotplug.
According to section 13.5.3 of the PAPR+, we have:
- set-power-level
- get-sensor-state
- set-indicator
- ibm,configure-connector (the RMO part is already covered in drmgr)

<disclaimer>
We should cover all the calls, but it requires a huge work and we
only want to enable PCI hotplug here. So yes, all other calls are
still broken in little endian... some more patching will be needed.
</disclaimer>

Signed-off-by: Greg Kurz <gkurz@linux.vnet.ibm.com>
Reviewed-by: Laurent Dufour <ldufour@linux.vnet.ibm.com>