librtasevent: Enable RTAS event parsing on LE systems
The librtasevent library does not work in LE environments. The
struct's for each rtas event section are defined as a bit field
to match layout of the rtas event. This librtasevent library
currently functions by mapping these structs directly the raw
rtas event that is passed. This works for BE systems because
rtas events are in BE by definition.

To fix this this patch re-defines the structs that have been
used to just contain an int for each value in the rtas event
sections, then adds a raw version of the struct that we can map
directly onto the passed in rtas event. The raw structs can then
be traversed an we can do any neccessary conversions to LE values
in the structs that are passed back to the user.

This patch updates most of the v6 rtas event sections, most notably
src/fru sections have not been updated yet. At this time there
are no plans to update pre-v6 rtas events for LE environments.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>