librtas: consolidate common actions in making a rtas call
Consolidate common actions, such as getting the rtas token and
handling extended delays, when making a rtas call into a routine.

This change presents to new ways to make a rtas call;
rtas_call() - Which handles any delays
rtas_call_no_delay() - Which does not handle any delays

The no delay variant was needed for the few instances where some work
needs to be done in between rtas calls when a delay occurs.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>