librtas: deprecate libofdt
The libofdt library is a project that was started a long time ago in
the hopes of providing a common infrastructure for accessing the device
tree. The code is mostly complete andd I really don't think there are
any users of it so I am proposing to deprecate the library.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>