reference compatibility dumps for abi-compliance-checker
Add XML dumps generated from the abi-compliance-checker tool against the
v2.0.3 tag using Ubuntu 20.04 and its GCC 9 cross toolchains. These dumps
capture the current binary and source-level interfaces in a format to be
used as input to the tool when checking an installed tree for
incompatibilities.

When we change the ABI or API, these dumps need to be regenerated.

Signed-off-by: Nathan Lynch <nathanl@linux.ibm.com>