avoid reserved names for inclusion guards in shipped headers
All identifiers that begin with an underscore and either an uppercase
letter or another underscore are reserved in the C standard.

The identifiers changed here are just inclusion guards and nothing
should be referring to them directly. Alter them to reflect the header
file names, adding corresponding comments to the #endif directives
where missing.

Signed-off-by: Nathan Lynch <nathanl@linux.ibm.com>
Fixes #4
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>