avoid reserved names for inclusion guards in internal header
All identifiers that begin with an underscore and either an uppercase
letter or another underscore are reserved in the C standard.

The identifier changed here is just an inclusion guard and nothing
should be referring to it directly. Alter it to reflect the header
file name, adding a corresponding comment to the #endif that
terminates the guard.

Signed-off-by: Nathan Lynch <nathanl@linux.ibm.com>
Fixes #4
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>