ensure non-powerpc builds do not attempt a real syscall
While we allow non-powerpc builds for the sake of coverage, the fact
that we force __NR_rtas to be defined means that such builds will
generate calls to nonsensical system calls. For example, syscall 255
is inotify_rm_watch on x86_64.

We don't anticipate anyone actually running non-powerpc librtas builds
to the point of attempting the rtas syscall, but it's best to ensure
that the syscall() invocation here always fails in an easily
understood way.

* Drop the local definition of __NR_rtas.

* When targeting powerpc, use SYS_rtas, which should always be
  defined.

* When targeting other architectures, use an invalid syscall number to
  ensure ENOSYS at runtime.

Signed-off-by: Nathan Lynch <nathanl@linux.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>