add ABI definitions as of v2.0.4
Generated against a v2.0.4 checkout using tools/generate-abi-defs.

Note: these include many visible symbols that are clearly intended for
internal use since they do not have declarations in the shipped
headers: basically every symbol without "rtas" in the name. The
question of whether these can be hidden in future releases is not
addressed in this change.

Future changes to the ABIs of librtas and librtasevent should be
accompanied by updates to these definitions.

Signed-off-by: Nathan Lynch <nathanl@linux.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>