add tools/generate-abi-defs script for ABI maintenance
This tool is invoked manually outside of the build system to generate
ABI definitions using the libabigail project's abidw tool:

https://sourceware.org/libabigail/manual/libabigail-overview.html

ABI definitions are generated in a Debian stable/bookworm container
for each of the PowerPC targets that the project has supported over
its life.

About half of the script is code generated by the argbash utility for
handling command line arguments.

https://argbash.readthedocs.io/en/latest/

Signed-off-by: Nathan Lynch <nathanl@linux.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>