build system: add support for cmocka-based tests
Add an --enable-tests switch to configure, which is off by
default. When enabled, check for a recent version of
cmocka (https://cmocka.org/) libraries to build tests with.

It is intended that 'configure --enable-tests' without a suitable
cmocka installation results in an error.

For now, the set of tests is empty.

Signed-off-by: Nathan Lynch <nathanl@linux.ibm.com>
Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>