librtas/vpd: prefer /dev/papr-vpd when available
Change rtas_get_vpd() to prefer the /dev/papr-vpd character device as
proposed for the kernel here:

https://lore.kernel.org/linuxppc-dev/20230822-papr-sys_rtas-vs-lockdown-v1-0-932623cf3c7b@linux.ibm.com/

On the first invocation of rtas_get_vpd(), probe for the new ABI and
initialize an internal function pointer according to the result. Use
pthread_once() to ensure that this probe step is performed atomically
and only once. It would not make sense to re-probe (and potentially
update the implementation) on every call, since the implementations
differ in the way state for the call sequence is maintained.

Assuming the new ABI is available: upon first call to rtas_get_vpd(),
open /dev/papr-vpd and obtain a file descriptor using the
PAPR_VPD_CREATE_HANDLE ioctl() command. This file descriptor is a
reference to the VPD returned from a complete ibm,get-vpd call
sequence. Subsequent calls sequentially read the VPD from the fd,
which is closed once EOF is reached.

To existing users of rtas_get_vpd(), the new implementation is
indistinguishable from the old one, with one exception:

* When using the rtas() syscall on current systems, RTAS advances the
  sequence number by a fixed interval on each call in a sequence. This
  is visible in the values returned in the 'seq_next' out parameter.

* When using /dev/papr-vpd, the value returned in 'seq_next' does not
  change between calls; librtas treats it as a "handle" for a sequence
  in progress.

In PAPR there is no meaning attached to the value returned in the
'seq_next' out parameter, and there is no requirement that it must
change or increment on each call. No user should be relying on the
behavior of this aspect of the interface to judge the status of a
sequence.

I have verified that an unmodified vpdupdate command from a distro
lsvpd package works correctly when linked to a librtas.so with this
change. This is the primary (and perhaps only) user of rtas_get_vpd().

For now, carry a copy of the kernel uapi header.

Signed-off-by: Nathan Lynch <nathanl@linux.ibm.com>