CI: relax -Werror to -Wall for clang builds
The clang compiler has recently generated format specifier warnings that
on first glance seem to be inconsistent with the C standard, and imply
unsigned types being promoted to sign types when passed as args to
printf. The result is the with -Werror these warnings are promoted to
errors and fail numerous builds in the matrix.

Relax -Werror to -Wall for clang builds while continuing to remain
pedantic with gcc builds for the time being.

Signed-off-by: Tyrel Datwyler <tyreld@linux.ibm.com>